🚚 폴더, 파일 네이밍 수정 및 경로 수정
