+++ frontend/src/constants/pairRoomModalInfo.ts
